Rename domain transfer events
