+++ docs/classes/Automattic-Domain-Services-Client-Event-Domain-Transfer-Out-Rejected.md
@@ -1,12 +1,13 @@
-# Class: [\Automattic](../namespaces/automattic.md)[\Domain_Services_Client](../namespaces/automattic-domain-services-client.md)[\Event](../namespaces/automattic-domain-services-client-event.md)[\Domain](../namespaces/automattic-domain-services-client-event-domain.md)[\Transfer](../namespaces/automattic-domain-services-client-event-domain-transfer.md)[\In](../namespaces/automattic-domain-services-client-event-domain-transfer-in.md)\Success
+# Class: [\Automattic](../namespaces/automattic.md)[\Domain_Services_Client](../namespaces/automattic-domain-services-client.md)[\Event](../namespaces/automattic-domain-services-client-event.md)[\Domain](../namespaces/automattic-domain-services-client-event-domain.md)[\Transfer](../namespaces/automattic-domain-services-client-event-domain-transfer.md)[\Out](../namespaces/automattic-domain-services-client-event-domain-transfer-out.md)\Rejected
 
 ## Summary:
 
-Inbound domain transfer success event
+Outbound domain transfer failure event
 
 ## Description:
 
-This event is generated when a domain transfer from another registrar to the reseller's account is successful.
+- This event is generated when a domain transfer to another registrar fails after it was successfully started
+- There might be more information about the cause of the failure in the event data
 
 
 ---
@@ -24,13 +25,12 @@ This event is generated when a domain transfer from another registrar to the res
 * public [get_request_date()](#method_get_request_date)
 * public [get_requesting_registrar()](#method_get_requesting_registrar)
 * public [get_transfer_status()](#method_get_transfer_status)
-* public [get_transferlocked_until_date()](#method_get_transferlocked_until_date)
 
 ---
 
 ### Details
 
-* File: [lib/event/domain/transfer/in/success.php](../../lib/event/domain/transfer/in/success.php)
+* File: [lib/event/domain/transfer/out/rejected.php](../../lib/event/domain/transfer/out/rejected.php)
 * Implements:
   * [\Automattic\Domain_Services_Client\Event\Event_Interface](../classes/Automattic-Domain-Services-Client-Event-Event-Interface.md)
   * [\Automattic\Domain_Services_Client\Event\Async_Command_Related_Interface](../classes/Automattic-Domain-Services-Client-Event-Async-Command-Related-Interface.md)
@@ -38,7 +38,6 @@ This event is generated when a domain transfer from another registrar to the res
   * [\Automattic\Domain_Services_Client\Event\Async_Command_Related_Trait](../classes/Automattic-Domain-Services-Client-Event-Async-Command-Related-Trait.md)
   * [\Automattic\Domain_Services_Client\Event\Object_Type_Domain_Trait](../classes/Automattic-Domain-Services-Client-Event-Object-Type-Domain-Trait.md)
   * [\Automattic\Domain_Services_Client\Event\Transfer_Trait](../classes/Automattic-Domain-Services-Client-Event-Transfer-Trait.md)
-  * [\Automattic\Domain_Services_Client\Event\TransferLocked_Trait](../classes/Automattic-Domain-Services-Client-Event-TransferLocked-Trait.md)
 
 ---
 
@@ -249,23 +248,4 @@ Gets the status of the transfer.
 
 ```
 string|null
-```
-
----
-
-<a id="method_get_transferlocked_until_date"></a>
-### get_transferlocked_until_date
-
-```
-public get_transferlocked_until_date() : \DateTimeInterface|null
-```
-
-##### Summary
-
-Gets the date until when the domain is transfer locked
-
-##### Returns:
-
-```
-\DateTimeInterface|null
 ```
