+++ docs/namespaces/automattic-domain-services-client-event-domain-transfer-out.md
@@ -4,6 +4,6 @@
 
 | Name | Summary |
 |------|---------|
-| [\Automattic\Domain_Services_Client\Event\Domain\Transfer\Out\Fail](../classes/Automattic-Domain-Services-Client-Event-Domain-Transfer-Out-Fail.md) | Outbound domain transfer failure event |
+| [\Automattic\Domain_Services_Client\Event\Domain\Transfer\Out\Completed](../classes/Automattic-Domain-Services-Client-Event-Domain-Transfer-Out-Completed.md) | Outbound domain transfer success event |
 | [\Automattic\Domain_Services_Client\Event\Domain\Transfer\Out\Pending](../classes/Automattic-Domain-Services-Client-Event-Domain-Transfer-Out-Pending.md) | Outbound domain transfer start event |
-| [\Automattic\Domain_Services_Client\Event\Domain\Transfer\Out\Success](../classes/Automattic-Domain-Services-Client-Event-Domain-Transfer-Out-Success.md) | Outbound domain transfer success event |
+| [\Automattic\Domain_Services_Client\Event\Domain\Transfer\Out\Rejected](../classes/Automattic-Domain-Services-Client-Event-Domain-Transfer-Out-Rejected.md) | Outbound domain transfer failure event |
